Rename stack command 'validate' to 'check'
